requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
supabase==2.0.3

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
html5lib==1.1

# Data processing
python-dateutil==2.8.2

# Development
pytest==7.4.3
httpx>=0.24.0,<0.25.0
